US117 Refactor Dashboard and Publications components
